+++ resources/migrations/20210317070506-user-secrets.up.sql
@@ -1,0 +1,5 @@
+CREATE TABLE user_secrets (
+  userId varchar(255) NOT NULL PRIMARY KEY,
+  secrets jsonb,
+  FOREIGN KEY (userId) REFERENCES users
+);